Move `garden` controllers / `garden-care` controller nits
