+++ pkg/operator/controller/add.go
@@ -30,8 +30,8 @@ import (
 	sharedcomponent "github.com/gardener/gardener/pkg/component/shared"
 	"github.com/gardener/gardener/pkg/controller/service"
 	"github.com/gardener/gardener/pkg/operator/apis/config"
-	"github.com/gardener/gardener/pkg/operator/controller/care"
-	"github.com/gardener/gardener/pkg/operator/controller/garden"
+	"github.com/gardener/gardener/pkg/operator/controller/garden/care"
+	"github.com/gardener/gardener/pkg/operator/controller/garden/garden"
 	"github.com/gardener/gardener/pkg/operator/controller/networkpolicyregistrar"
 	gardenerutils "github.com/gardener/gardener/pkg/utils/gardener"
 	"github.com/gardener/gardener/pkg/utils/imagevector"
