+++ pkg/operator/controller/garden/care/add_test.go
@@ -31,7 +31,7 @@ import (
 	gardencorev1beta1 "github.com/gardener/gardener/pkg/apis/core/v1beta1"
 	operatorv1alpha1 "github.com/gardener/gardener/pkg/apis/operator/v1alpha1"
 	operatorclient "github.com/gardener/gardener/pkg/operator/client"
-	. "github.com/gardener/gardener/pkg/operator/controller/care"
+	. "github.com/gardener/gardener/pkg/operator/controller/garden/care"
 )
 
 var _ = Describe("Add", func() {
