+++ pkg/operator/controller/garden/care/care_suite_test.go
@@ -23,5 +23,5 @@ import (
 
 func TestCare(t *testing.T) {
 	RegisterFailHandler(Fail)
-	RunSpecs(t, "Garden Controller Care Suite")
+	RunSpecs(t, "Operator Controller Garden Care Suite")
 }
