+++ pkg/operator/controller/garden/garden/reconciler_delete.go
