+++ skaffold-operator.yaml
@@ -153,8 +153,8 @@ build:
         - pkg/operator/apis/config/validation
         - pkg/operator/client
         - pkg/operator/controller
-        - pkg/operator/controller/care
-        - pkg/operator/controller/garden
+        - pkg/operator/controller/garden/care
+        - pkg/operator/controller/garden/garden
         - pkg/operator/controller/networkpolicyregistrar
         - pkg/operator/features
         - pkg/operator/webhook
